Finished i2c-proc.c
When you insert i2c-proc.o, you will find some files in /proc/bus (which
is first created for 2.0 kernels). Currently, the following files are
implemented:
  * /proc/bus/i2c
    A list of i2c/smbus adapters. If DEBUG was true when you compiled this
    module, you will also see the ISA pseudo-adapter (if isa.o is loaded).
  * /proc/bus/i2c-?
    For each adapter in /proc/bus/i2c, a list of registered clients.
    If DEBUG was true when you compiled this module, you will also see
    the i2c-proc pseudo client.
Of course, these entries are modified when adapters are added/removed.

Kernel 2.1.x users are encouraged to test this module, as /proc entries
are somewhat different beginning with kernel 2.1.29.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@18 7894878c-1315-0410-8ee3-d5d059ff63e0