Expanded expressions, slightly new syntax
An expression can now refer to the value it has to compute (@), but also to
other values. This is now perfectly valid as an expression:
  vid + in_min - in_max * 2 / @
Not that this is very sensible, but you get the idea. Please install the
new example config file, or replace all old 'val' occurences with '@'.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@107 7894878c-1315-0410-8ee3-d5d059ff63e0