(Phil) Added a driver for the Analog Devices ADM1021.  The ADM1021 is basically an expanded LM75.  It can have an external temp probe (in addition to the internal one), customizable conversion rate (in # of times per second, depending on power consumption requirements), and actually has a device-id register for easier/more-certain detecting!
The only strange property is the low-limit.  The low-limit doesn't act
as a Hystersis value like the LM75.  I.e., if the temp falls below it,
it triggers an alarm instead of clearing it.  Strange, ey?  This makes
using the chip for a thermostat control (like for a fan) much harder.
But it *could* be used to turn on a heater to prevent your computer from
freezing. :')

Here's what my output looks like (without a remote sensor. temp's
are the same format as the LM75 [high-limit, low-limit, value]):

[root@penn adm1021-i2c-0-18]# ls -l
total 0
-r--r--r--   1 root     root            0 Jan  9 23:42 die_code
-rw-r--r--   1 root     root            0 Jan  9 23:42 remote_temp
-r--r--r--   1 root     root            0 Jan  9 23:42 status
-rw-r--r--   1 root     root            0 Jan  9 23:42 temp
[root@penn adm1021-i2c-0-18]# cat *
3
60 20 0
76
60 20 24


PS- all temp values in the ADM1021 (limits and readings) are signed
(2's comp) bytes.  So, all readings go from -128 to 127 C (actually,
the docs say it can only reliable sense down to -65), and only
in whole degree increments.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@153 7894878c-1315-0410-8ee3-d5d059ff63e0