(Phil) Added a driver for the LTC1710.  The LTC1710 is a very simple dual switch.  It can switch two devices which each consume 300mA at 5V (or less). This is good for fans, lights, or into relays for heftier stuff.
Here's what the usage looks like:

[root@penn sensors]# ls ltc1710-i2c-1-58/*
ltc1710-i2c-1-58/switch1
ltc1710-i2c-1-58/switch2

[root@penn sensors]# cat ltc1710-i2c-1-58/*
0
0

[root@penn sensors]# echo "1" >  ltc1710-i2c-1-58/switch2
(blue LED turned on)

[root@penn sensors]# cat ltc1710-i2c-1-58/*
0
1

One problem with the device: It doesn't support reads.  It ignores any
read transactions, so even detection with a read won't work.  This
also means that the status of the switches has to be remembered and
can't be polled.

As simple as this device is, it is one of the few which actually
allows the computer to send signals OUT into the world to
control something.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@155 7894878c-1315-0410-8ee3-d5d059ff63e0