(Phil) Oh boy, here we go! :')  This is the removal of the original files found under lm_sensors2/src and moved to lm_sensors/kernel (and new sub-directories), as suggested by Frodo (from his email):
  kernel (i2c-dev.c, i2c-proc.c, sensors.c, smbus.c, compat.h)
   |
   |-- include (smbus.h, sensors.h, i2c-dev.h, isa.h)
   |
   |-- busses (i2c-isa.c, i2c-piix4.c, i2c-via.c)
   |
   |-- chips (adm1021.c, adm9240.c, eeprom.c, gl518sm.c, lm75.c, lm78.c,
              lm80.c, ltc1710.c, w83781d.c)

There may need to be more shuffling of a few more files, but this is the
bulk of the movement.  Of course, none of this will compile without some
changes to the appropriate 'make' files.

Also note that this commit notes only the removal and not the insertion
of the copies because that insertions were done by hand on the CVS
repository its self.

Last note, the change history (and comment logs) are intact, but the tags
were removed on the new files (because the files didn't really exist when
the tags were first made).

Don't forget to do a 'cvs update -d', and to also remove lm_sensors2/src!


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@175 7894878c-1315-0410-8ee3-d5d059ff63e0