Update to the latest (mod-19990118) kernel archive
I really hope I did not forget an add or remove. If strange things happen
when you compile, please let me know.

There may still be problems with the EXPORT_* stuff; it seems to work now
for kernel 2.0.36pre19, but it may give trouble for earlier kernels. We
need to check that.

Some i2c modules may be a bit verbose at this moment.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@184 7894878c-1315-0410-8ee3-d5d059ff63e0