Documentation updates, small bug fixes
I am trying to create a more-or-less standard documentation format for
doc/chips/* files. I am not quite satisfied with the lm75 and lm78 docs
yet, but it is starting to come. Basically, I want as much information
in it as possible, and to generate most of it automatically. But that is
harder than I though it would be.

Also, a small library bugfix (class problem) and a doc-features fix.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@210 7894878c-1315-0410-8ee3-d5d059ff63e0