Changed forcibly enabled stuff to insmod parameters.
There are now two new insmod parameters, and the #define has disappeared

force: if set to anything but zero, this enables the PIIX4 as the old
       #define did.
force_addr: This first sets a new I/O port address for the PIIX4, and
       afterwards enables it. Supercedes 'force'.

Note that using force_addr is very, very dangerous if you do not know
what you are doing. On the other hand, it was very easy to add, and can
be a big help if your BIOS doesn't set correct addresses.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@353 7894878c-1315-0410-8ee3-d5d059ff63e0