Rewrote the doc-features program in Perl.
This program scan the C-sources of the library and all kernel modules and
outputs everything it can find about all features.

To do: sort features and add command-line parameters, implement slightly
       better layout.

The perl program completely replaces the C program and already has
much more functionality.

This program will be very helpful to write the kernel module documentation,
and especially to keep it up-to-date: as the C sources are examined,
updates need only be done at one place, in the source code.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@384 7894878c-1315-0410-8ee3-d5d059ff63e0