(Phil) This is a simple Perl script which reads standard input and sends the results to the display.  Only the first four lines are read, and only the first 20 chars of each line are displayed.
Example use:

echo -e "hi\nthere,\nhow are\nyou?" | displayit.pl

Displays this on the display:

hi
there,
how are
you?


It doesn't get much simpler than this!

TODO: The code isn't very efficient and goes slow for what it is doing.
It assumes that the display is 20x4 (easily changable).  And, it assumes
that there is only one display. This is really more of a simple example
than a feature-rich app.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@483 7894878c-1315-0410-8ee3-d5d059ff63e0