Added some warnings in sensors-detect
It now warns you about built-in drivers, and should always work. It will
still give a too long list of modules if some things are built into the
kernel. There is no real solution to this; the best thing to do would
probably be to ask the user where the kernel tree is located, and use
.config to determine what is built into the kernel. But not everybody
has such a kernel tree - especially once we get to the point where
people will distribute binary versions.

I also updated the list of undetectable adapter.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@589 7894878c-1315-0410-8ee3-d5d059ff63e0