Apply patch from David Gesswein:
The following patch does the following:

Fixed bug with not all alarms enabled.
Added ability to read battery voltage
Select temperature sensor type at module load time.
Added setting for Inside Technologies 786LCD board to example config file
I also fixed what looked like a typo in the existing it87 section where
in6_min/max was set twice and in7 was not set.  As commented I think
the equation for the negative voltages are wrong but without that motherboard
I can't verify.

With the 786LCD board sensors-detect reports a lm78 which doesn't
exist and a sis5595 which may be part of the SIS630 chip but doesn't
perform the monitoring.  The it8705F/SIS950 which is on the board is
reported as a misdetect.  I didn't know enough about the various chips
to figure out how to get correct detections.

David Gesswein


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@1146 7894878c-1315-0410-8ee3-d5d059ff63e0