(mds)         Subject:         Uptodate i2c-keywest    Date:         Sun, 17 Feb 2002 21:34:44 +0100   From:         Benjamin Herrenschmidt <benh@kernel.crashing.org>
Here's a patch against current lmsensors CVS which updates i2c-keywest
to the latest version needed on powermac. It's already in Marcelo as
one of the 2.4.18-pre's.

You may note that the i2c "mode" isn't what it should be to match
exactly what linux-i2c expects (see explaining comment). I will have
to define new transfer types in order to work around that, but it's
ok to do that now as we use this driver only for the sound chip
for now.

Ben.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@1326 7894878c-1315-0410-8ee3-d5d059ff63e0