Patch by Kris Van Hees:
Here is a small patch that fixes the setting of limits on the temp2 sensors
in the MTP008 chip.  The code (embarassingly written by myself when I wrote
the MTP008 chip driver a long time ago) would always update the hyst and max
limits of the first temp sensor, even when you tried to set it for the 2nd
or 3rd sensors.


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@2649 7894878c-1315-0410-8ee3-d5d059ff63e0