Patch from Anton V. Staaf:
"(...) I found what I think is a bug in the it87_fan_div setting code.
(...) What was happening was that the check to make sure the third fan's
divisor was either 2 or 8 was using the wrong member of the it87 data
structure.  This was causing the if to never fail and thus always force
the divisor to 2."


git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@3200 7894878c-1315-0410-8ee3-d5d059ff63e0