Modify libsensors scanner build flags...
* add -8
When using -Cf, flex defaults to generating a 7-bit scanner instead of 8-bit.
That's a bug, since we want the scanner to handle any character.  Using -8 restores
the proper behavior.

* change -Cf to -Cfe
This allows a nice reduction in object size (from ~60KiB to ~15KiB) with almost
no impact to speed (adds < 100 cycles to scanner, where total cycles is > 1M).



git-svn-id: http://lm-sensors.org/svn/lm-sensors/trunk@4272 7894878c-1315-0410-8ee3-d5d059ff63e0