fancontrol: Reduce memory consumption
Construct "wait $!" causes bash to remember the exit status of the
children. This consumes memory. We don't need the status code and we
are only waiting for a single child, so we can simply use "wait". This
will reduce the memory usage.

This partly fixes ticket #2391.
http://www.lm-sensors.org/ticket/2391

The actual leak is in bash 4.3 and should be fixed there.