Add 'configs' directory where sample configurations can be stored
Most mainboards need a config file which describe which input of a
monitoring chip is connected to which voltage rail, what are scaling
factors of voltages that can't be measured directly by the chip, which
part of the board has each temperature sensor, which fan is connected to
which tachometer input of the chip, etc.

All these data need to be determined either from a board schematic (if
available) or by a trial-and-error method.

This can be a time consuming and error-prone work, fortunately, it needs to
be done just once per mainboard.
Currently, however, there is no standard mechanism to share such
configurations so multiple users of the same board don't basically waste
their time reinventing the wheel (or recreating the same configuration
again and again).

Previously, a wiki on an old lm-sensors site had a 'Configurations' page
where contributors could submit their configurations.

Since this site is down for two years now and it does not look like it is
coming back we need to find an another solution.

Let's use a 'configs' directory in project repository for these files, so
we can benefit from SCM infrastructure for managing them.
Also, add a reminder in documentation asking people to contribute these
files for benefit of the whole community and configurations for 3
mainboards for a starter.

Signed-off-by: Maciej S. Szmigiero <mail@maciej.szmigiero.name>