The original calculation did not work, the pwm was always set to maximum.
The formular for calculating the new pwm value is redifined as i could not comprehend why it was this way. Generally, integer operations are not a good idea at that point, even if input and output are integer based calculations including divisions are better done in float. At least for my fan here this change worked perfecty as expected. It introduces of course a new dependence on bc.