fix wrong alg description, + intention of use
The description of the algorithm was wrong, this is fixed. it also reflects a new proposed variable in the config for turning the linear into a square transfer function. 
The chapter of description was only a functional description of what is does. An intention of use is in my opinion fundamentally important for users.