configs: Bulk config recovery from archive.org (Part 2)
This is a bulk recovery of the sensor configs found at
https://web.archive.org/web/20150901092438/http://www.lm-sensors.org:80/wiki/Configurations

The config sections themselves are unaltered, but I've selectively added
some, but not all of the extra info around them. Anywhere there's been
extra info added, it should be in comments at the top of the config,
followed by a double newline to indicate the separation. Attribution has
been retained where possible, but links to the mailing list are hosed,
so goto the above archive.org link to check if a config has extra origin
notes.

[groeck: reformatted description]
Signed-off-by: Guenter Roeck <linux@roeck-us.net>