sensors: json: Do not print a stray comma
Previously, if an error happened when reading the value of a subfeature,
subCnt would get incremented and a stray comma would get printed in the
following iteration, even though nothing was printed in the previous
iteration. This would produce a syntactically incorrect JSON.

This is based on a patch from the su8 user on GitHub.

Signed-off-by: Ondřej Lysoněk <olysonek@redhat.com>