Replace svn keywords with LM_SENSORS_VER constant
The keywords Revision, Date and Id were being expanded by Subversion
way back when the project was being maintained in Subversion. Now when
the project is maintained in git, the keywords are no longer expanded
and appear e.g. in the output of sensors-detect literally.

Drop the keywords, or replace them with LM_SENSORS_VER where
applicable.

The value of the constant will have to be updated manually every time
a new release is made.

Inspired by a similar change in i2c-tools (commit daa430031ffd;
thanks, Jean).