sensors-detect: add support for FTS Teutates Chip
This patch adds support for the FTS Teutates Chip,
used on Fujitsu Mainboards for Intel Skylake and later Chipsets.
It is based on a patch from Fujitsu and was tested on a D-3446-S Mainboard.

The corresponding kernel module can be found here [1].

This is a revised version of an earlier pull request [2].

[1]: https://github.com/torvalds/linux/blob/master/drivers/hwmon/ftsteutates.c#L711
[2]: https://github.com/lm-sensors/lm-sensors/pull/9