silence sensors -u stderr
Normal output of sensors and sensors -u differs also in the amount of potential error messages sent to stderr.
The -u option is meant for post processing. I suggest to keep the error messages inline with the normal print on stdout, so that it is possible to derive which subvalue had the error.
Post processing currently has no way to correlate the error message with the subvalues.