sensors: Handle json trailing ',' when no features are listed
The json format requires that the last entry of a dictionary does not
have a trailing ','. To avoid generating invalid json output, check that
there are features to list before emitting the trailing ','.

This issue is present when a chip exists but all its sensors are
ignored:
{
    "example-device" : {
      "Adapter": "example adapter",
    }
}

The corrected output is:
{
    "example-device" : {
      "Adapter": "example adapter"
    }
}

Signed-off-by: Nathan Rossi <nathan.rossi@digi.com>